+++ src/Presentation/SmartStore.Web.Framework/UI/Captcha/HtmlExtensions.cs
@@ -1,28 +1,30 @@
-﻿using System.IO;
+﻿using System.Text;
 using System.Web.Mvc;
-using System.Web.UI;
 using SmartStore.Core.Infrastructure;
+using SmartStore.Utilities;
 
 namespace SmartStore.Web.Framework.UI.Captcha
 {
-    public static class HtmlExtensions
+	public static class HtmlExtensions
     {
         public static string GenerateCaptcha(this HtmlHelper helper)
         {
+			var sb = new StringBuilder();
             var captchaSettings = EngineContext.Current.Resolve<CaptchaSettings>();
-            var captchaControl = new Recaptcha.RecaptchaControl
-            {
-                ID = "recaptcha",
-                Theme = "white",
-                PublicKey = captchaSettings.ReCaptchaPublicKey,
-                PrivateKey = captchaSettings.ReCaptchaPrivateKey
-            };
+			var widgetUrl = CommonHelper.GetAppSetting<string>("g:RecaptchaWidgetUrl");
+			var elementId = "GoogleRecaptchaWidget";
 
-            var htmlWriter = new HtmlTextWriter(new StringWriter());
+			sb.AppendLine("<script type=\"text/javascript\">");
+			sb.AppendLine("var googleRecaptchaOnloadCallback = function() {");
+			sb.AppendLine("  grecaptcha.render('{0}', {{".FormatInvariant(elementId));
+			sb.AppendLine("    'sitekey' : '{0}'".FormatInvariant(captchaSettings.ReCaptchaPublicKey));
+			sb.AppendLine("  });");
+			sb.AppendLine("};");
+			sb.AppendLine("</script>");
+			sb.AppendLine("<div id=\"{0}\"></div>".FormatInvariant(elementId));
+			sb.AppendLine("<script src=\"{0}?onload=googleRecaptchaOnloadCallback&render=explicit\" async defer></script>".FormatInvariant(widgetUrl));
 
-            captchaControl.RenderControl(htmlWriter);
-
-            return htmlWriter.InnerWriter.ToString();
+			return sb.ToString();
         }
     }
 }