+++ changelog.md
@@ -83,6 +83,7 @@
 * Trusted Shops: badge will be displayed in mobile themes, payment info link replaced compare list link in footer
 * Product filter: Specification attributes are sorted by display order rather than alphabetically by name
 * #856 Don't route topics which are excluded from sitemap
+* #851 Replace reCAPTCHA with "I'm not a robot" CAPTCHA
 
 ### Bugfixes
 * #523 Redirecting to payment provider performed by core instead of plugin